+++ src/Formio.js
@@ -560,8 +560,7 @@ export default class Formio {
     return NativePromise.all([
       (form !== undefined) ? NativePromise.resolve(form) : this.loadForm(),
       (user !== undefined) ? NativePromise.resolve(user) : this.currentUser(),
-      // load submission only if it's not passed directly and if submissionId is defined
-      (!submission && this.submissionId) ? this.loadSubmission() : NativePromise.resolve(submission),
+      (submission !== undefined || !this.submissionId) ? NativePromise.resolve(submission) : this.loadSubmission(),
       this.accessInfo()
     ]).then((results) => {
       const form = results.shift();